% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{power.mvnTest}
\alias{power.mvnTest}
\title{Power Calculation using the Zhou-Shao's Multivariate Normality Test Statistic (\eqn{T_n})}
\usage{
power.mvnTest(a, n, p, B = 1000, pct = c(0.01, 0.99), FUN, ...)
}
\arguments{
\item{a}{significance level (\eqn{\alpha}).}

\item{n}{number of rows (observations).}

\item{p}{number of columns (variables), \eqn{n>p}.}

\item{B}{number of Monte Carlo simulations, default is 1000 (can increase B to increase the precision).}

\item{pct}{percentiles of MK to get c1 and c2 described in the reference paper,default is (0.01, 0.99).}

\item{FUN}{self-defined function for generate multivariate distribution. See example.}

\item{...}{optional arguments passed to \code{FUN}.}
}
\value{
Returns a numeric value of the estimated empirical power (value between 0 and 1).
}
\description{
Empirical power calculation using the Zhou-Shao's multivariate normality test Statistic \eqn{T_n}.
}
\examples{
set.seed(12345)

## Power calculation against bivariate (p=2) independent Beta(1, 1) distribution ##
## at sample size n=50 for Tn at one-sided alpha = 0.05 ##

power.mvnTest(a = 0.05, n = 50, p = 2,  B = 100, pct = c(0.01, 0.99), FUN=IMMV, D1=runif)

}
\references{
Zhou, M., & Shao, Y. (2014). A powerful test for multivariate normality. \emph{Journal of applied statistics}, 41(2), 351-363.
}
